XML changes for SLES 11 vs 12, and gpt vs msdos :black_circle:
This should hopefully get all the cases to get autoyast happy. When using normal partitioning the msdos would kill the existing labels. So that required the advanced_partitioning. The advanced partitioning did not work in Joe's environment with sles 12 gpt. Not sure why, but normal partitioning did work there, so had to go back to normal partitioning for that one.