When using shortnames, stack run host can break
When an interface in the database is set to use shortnames, that
becomes its quasi-default interface, since that shortname is how
run host connects to the node.

However, it's possible that this interface is not on a network that
the frontend can access. If that's the case, the frontend cannot run
a run host command against it.

To avoid this problem, we use the 'stack.network' attribute. This will
tell 'run host' which network to use when talking to hosts. If this
attribute doesn't exist, 'run host' will use the shortname
(default old-style behaviour)